requirements-dev.txt
-r requirements.txt
isort==5.13.2
pytest==8.2.1
requests-mock==1.12.1
ruff==0.4.4
